pyproject.toml
[project]
name = "skate-spots"
version = "0.1.0"
description = "A Django application for sharing and discovering skateboarding spots"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "Django>=5.0,<5.1",
    "djangorestframework>=3.14.0",
    "djangorestframework-simplejwt>=5.3.0",
    "django-environ>=0.11.0",
    "django-ratelimit>=4.1.0",
    "django-filter>=23.0",
    "drf-spectacular>=0.27.0",
    "structlog>=24.3.0",
    "python-json-logger>=2.0.0",
    "python-multipart>=0.0.20",
    "bcrypt>=4.1.2,<5.0",
    "email-validator>=2.3.0",
]


[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
]
ignore = [
    "E501",  # line too long (handled by line-length setting)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "ARG002",  # Unused method argument (fixtures are intentional)
    "B017",    # Blind exception assertion (acceptable in tests)
]
"*/migrations/*.py" = [
    "E501",    # Line too long (Django migrations can be long)
]
"accounts/views.py" = [
    "ARG002",  # Unused method arguments (intentional in Django ViewSets)
]
"spots/views.py" = [
    "ARG",     # Unused method arguments (intentional in Django ViewSets and actions)
]
"spots/filters.py" = [
    "ARG002",  # Unused name argument (intentional in django-filter)
]
"spots/permissions.py" = [
    "ARG002",  # Unused view argument (intentional in DRF permissions)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_tests.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "--disable-warnings",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests",
]

[tool.django-stubs]
django_settings_module = "skate_spots_project.settings"

[dependency-groups]
dev = [
    "ruff>=0.12.10",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-django>=4.7.0",
]
